+++ R/matrix_eval.R
@@ -18,13 +18,15 @@
 check_matrix <- function(x, ...) {
   info <- list(...)
   
-  stopifnot(
-    isTRUE(
-      all.equal(rowSums(x), rep(1, nrow(x)))
-    ),
-    all(x >= 0 & x <= 1)
-  )
+  if (! isTRUE(all.equal(rowSums(x), rep(1, nrow(x))))) {
+    stop("Not all transition matrix rows sum to 1.")
+  }
+  
+  if (! all(x >= 0 & x <= 1)) {
+    stop("Some transition probabilities are outside the interval [0 - 1].")
+  }
 }
+
 
 #' Evaluate Markov Model Transition Matrix
 #' 
@@ -37,43 +39,48 @@ check_matrix <- function(x, ...) {
 #'   
 #' @return An \code{eval_matrix} object (actually a list of 
 #'   transition matrix, one per cycle).
+
 eval_matrix <- function(x, parameters) {
-  
+
   tab_res <- mutate_(parameters, C = -pi, .dots = x)[names(x)]
   
   n <- get_matrix_order(x)
-  
-  f <- function(...) {
-    res <- matrix(c(...),
-                  byrow = TRUE,
-                  nrow = n)
-    posC <- res == -pi
+
+  f3 <- function(this_array) {
+    posC <- this_array == -pi
     
     stopifnot(
-      rowSums(posC) <= 1
+      all(rowSums(posC, dims = 2) <= 1)
     )
-    res[posC] <- 0
-    valC <- 1 - rowSums(res)[which(posC, arr.ind = TRUE)[, 1]]
-    res[posC] <- valC
+    this_array[posC] <- 0
     
-    check_matrix(res)
-    list(res)
+    valC <- 1 - rowSums(this_array, dims = 2)[which(posC, arr.ind = TRUE)[, -3]] 
+    this_array[posC] <- valC
+    
+    sapply(1:dim(this_array)[1], function(i){check_matrix(this_array[i,,])})
+    this_array
   }
   
-  # bottleneck!
+  tab_res_2 <- array(unlist(tab_res), dim = c(nrow(tab_res), n, n))
   
-  res <- unlist(
-    dplyr::do(
-      dplyr::rowwise(tab_res),
-      res = f(unlist(.))
-    )$res,
-    recursive = FALSE
-  )
+  for(i in 1:nrow(tab_res)){
+    tab_res_2[i,,] <- t(tab_res_2[i,,])
+  }
   
-  structure(res,
-            class = c("eval_matrix", class(res)),
+  tab_res_3 <- f3(tab_res_2)
+  tab_res_4 <- split_along_dim(tab_res_3, 1)
+  
+  structure(tab_res_4,
+            class = c("eval_matrix", class(tab_res_4)),
             state_names = get_state_names(x))
 }
+
+
+split_along_dim <- function(a, n)
+  setNames(lapply(split(a, arrayInd(seq_along(a), dim(a))[, n]),
+                  array, dim = dim(a)[-n], dimnames(a)[-n]),
+           dimnames(a)[[n]])
+
 
 get_state_names.eval_matrix <- function(x, ...){
   attr(x, "state_names")