+++ R/model_print.R
@@ -15,7 +15,7 @@ print.uneval_model <- function(x, ...) {
   ))
 }
 
-#' Plot Results of Markov Model
+#' Plot Results of a Markov Model
 #' 
 #' Various plots for Markov models.
 #' 
@@ -31,7 +31,7 @@ print.uneval_model <- function(x, ...) {
 #' @return A \code{ggplot2} object.
 #' @export
 #' 
-plot.eval_model_list <- function(x, type = c("counts", "ce"), model = 1, ...) {
+plot.run_models <- function(x, type = c("counts", "ce"), model = 1, ...) {
   type <- match.arg(type)
   
   switch(
@@ -69,6 +69,3 @@ plot.eval_model_list <- function(x, type = c("counts", "ce"), model = 1, ...) {
     stop(sprintf("Unknown type: '%s'.", type))
   )
 }
-if(getRversion() >= "2.15.1")
-  utils::globalVariables(c("row_number", "markov_cycle", "value", "key",
-                           ".cost", ".effect", ".model_names"))