+++ R/run_model_define.R
@@ -53,7 +53,7 @@ run_models <- function(...,
                        cycles = 1,
                        method = c("beginning", "end",
                                   "half-cycle", "life-table"),
-                       cost, effect, base_model = NULL) {
+                       cost = NULL, effect = NULL, base_model = NULL) {
   list_models <- list(...)
   
   method <- match.arg(method)
@@ -64,8 +64,8 @@ run_models <- function(...,
     init = init,
     cycles = cycles,
     method = method,
-    cost = lazyeval::lazy(cost),
-    effect = lazyeval::lazy(effect),
+    cost = lazyeval::lazy_(substitute(cost), env = parent.frame()),
+    effect = lazyeval::lazy_(substitute(effect), env = parent.frame()),
     base_model = base_model
   )
 }
@@ -79,18 +79,22 @@ run_models_ <- function(list_models,
                         method,
                         cost, effect, base_model) {
   
-  stopifnot(
-    all(unlist(lapply(list_models,
-                      function(x) "uneval_model" %in% class(x)))),
-    ! missing(cost),
-    ! missing(effect)
-  )
+  if (! all(unlist(lapply(list_models,
+                          function(x) "uneval_model" %in% class(x))))) {
+    .x <- names(list_models[! unlist(lapply(list_models,
+                                            function(x) "uneval_model" %in% class(x)))])
+    stop(sprintf(
+      "Incorrect model object%s: %s.",
+      plur(length(.x)),
+      paste(.x, collapse = ", ")
+    ))
+  }
   
   list_ce <- list(
-    cost,
-    effect
+    .cost = cost,
+    .effect = effect
   )
-  names(list_ce) <- c(".cost", ".effect")
+  
   ce <- c(
     lazyeval::lazy_dots(),
     list_ce
@@ -110,25 +114,34 @@ run_models_ <- function(list_models,
     names(list_models) <- model_names
   }
   
-  stopifnot(
-    all("uneval_model" %in% unlist(lapply(list_models, class))),
-    list_all_same(lapply(list_models,
-                         function(x) sort(get_state_names(x)))),
-    list_all_same(lapply(list_models,
-                         function(x) sort(get_state_value_names(x))))
-  )
+  if (! list_all_same(lapply(list_models,
+                             function(x) sort(get_state_names(x))))) {
+    stop("State names differ between models.")
+  }
   
-  stopifnot(
-    length(init) == get_state_number(list_models[[1]]),
-    all(init >= 0)
-  )
+  if (! list_all_same(lapply(list_models,
+                             function(x) sort(get_state_value_names(x))))) {
+    stop("State value names differ between models.")
+  }
+  
+  if (! length(init) == get_state_number(list_models[[1]])) {
+    stop(sprintf(
+      "Length of 'init' vector (%i) differs from number of states (%i).",
+      length(init),
+      get_state_number(list_models[[1]])
+    ))
+  }
+  
+  if (! any(init > 0)) {
+    stop("At least one init count must be > 0.")
+  }
   
   if (is.null(names(init)))
     names(init) <- get_state_names(list_models[[1]])
   
-  stopifnot(
-    all(sort(names(init)) == sort(get_state_names(list_models[[1]])))
-  )
+  if (! all(sort(names(init)) == sort(get_state_names(list_models[[1]])))) {
+    stop("Names of 'init' vector differ from state names.")
+  }
   
   eval_model_list <- lapply(list_models, eval_model, 
                             parameters = parameters,
@@ -142,9 +155,8 @@ run_models_ <- function(list_models,
     list_res[[n]]$.model_names <- n
   }
   
-  res <- Reduce(dplyr::bind_rows, list_res)
-  
-  res <- dplyr::mutate_(res, .dots = ce)
+  res <- Reduce(dplyr::bind_rows, list_res) %>% 
+    dplyr::mutate_(.dots = ce)
   
   if (is.null(base_model)) {
     base_model <- get_base_model(res)
@@ -154,7 +166,7 @@ run_models_ <- function(list_models,
     res,
     eval_model_list = eval_model_list,
     uneval_model_list = list_models,
-    class = c("eval_model_list", class(res)),
+    class = c("run_models", class(res)),
     parameters = parameters,
     init = init,
     cycles = cycles,
@@ -164,22 +176,12 @@ run_models_ <- function(list_models,
   )
 }
 
-#' Get Markov Model Parameters
-#' 
-#' 
-#' For internal use.
-#' 
-#' @param x An \code{eval_model_list}
-#'   object.
-#'   
-#' @return An \code{uneval_parameters} or
-#'   \code{eval_parameters} object.
-get_parameters <- function(x){
-  UseMethod("get_parameters")
+get_model_names <- function(x) {
+  x$.model_names
 }
 
-get_parameters.default <- function(x){
-  attr(x, "parameters")
+get_model_count <- function(x) {
+  nrow(x)
 }
 
 get_total_state_values <- function(x) {
@@ -195,9 +197,13 @@ get_base_model <- function(x, ...) {
 }
 
 get_base_model.default <- function(x, ...) {
-  x$.model_names[which(x$.effect == min(x$.effect))[1]]
+  if (! ".effect" %in% names(x)) {
+    warning("No effect defined, cannot find base model.")
+    return(NULL)
+  }
+  x$.model_names[x$.effect == min(x$.effect)][1]
 }
 
-get_base_model.eval_model_list <- function(x, ...) {
+get_base_model.run_models <- function(x, ...) {
   attr(x, "base_model")
 }