+++ R/sensitivity_eval.R
@@ -11,19 +11,29 @@
 #' @example inst/examples/example_run_sensitivity.R
 run_sensitivity <- function(model, sensitivity) {
   
+  if (! all(c(".cost", ".effect") %in% names(model))) {
+    stop("No cost and/or effect defined, sensitivity analysis unavailable.")
+  }
+  
   init <- attr(model, "init")
   cycles <- attr(model, "cycles")
   method <- attr(model, "method")
-  list_models <- attr(model, "uneval_model_list")
+  names_models <- get_model_names(model)
   
-  list_res <- lapply(list_models, eval_model_newdata,
-                     method = method, old_parameters = get_parameters(model),
-                     init = init, cycles = cycles, newdata = sensitivity)
-  for (n in names(list_res)) {
-    list_res[[n]]$.model_names <- n
+  list_res <- lapply(
+    names_models,
+    function(n) eval_model_newdata(model, model = n, newdata = sensitivity)
+  )
+  
+  for (i in seq_along(names_models)) {
+    list_res[[i]]$.model_names <- names_models[i]
   }
   
   res <- Reduce(dplyr::bind_rows, list_res)
+  res <- res %>% 
+    dplyr::rowwise() %>% 
+    dplyr::do(get_total_state_values(.$.mod)) %>% 
+    dplyr::bind_cols(res)
   
   res <- dplyr::mutate_(res, .dots = attr(model, "ce"))